--- src/main/java/com/traveloper/tourfinder/api/KTO/dto/area/AreaBasedDto.java
@@ -1,4 +1,0 @@
-package com.traveloper.tourfinder.api.KTO.dto.area;
-
-public class AreaBasedDto {
-}
