+++ src/main/java/com/traveloper/tourfinder/course/controller/PlaceController.java
@@ -1,21 +1,12 @@
 package com.traveloper.tourfinder.course.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.traveloper.tourfinder.api.KTO.dto.KTOKeywordSearchDto;
+import com.traveloper.tourfinder.api.KTO.dto.search.KTOKeywordSearchDto;
 import com.traveloper.tourfinder.api.KTO.dto.detail.DetailsCommonDto;
-import com.traveloper.tourfinder.api.KTO.dto.detail.DetailsItemDto;
 import com.traveloper.tourfinder.api.KTO.service.KTOApiService;
-import com.traveloper.tourfinder.course.dto.PlaceDto;
-import com.traveloper.tourfinder.course.entity.Place;
 import com.traveloper.tourfinder.course.service.PlaceService;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
